+++ shared/src/androidInstrumentedTest/kotlin/InstrumentedTests.android.kt
@@ -7,5 +7,5 @@ import kotlinx.coroutines.Dispatchers
 @Composable
 actual fun getPlatformAdapter(): PlatformAdapter {
     val context = LocalContext.current
-    return AndroidPlatformAdapter(context, {})
+    return AndroidPlatformAdapter(context)
 }