+++ shared/src/commonMain/kotlin/App.kt
@@ -4,7 +4,7 @@ import androidx.compose.ui.graphics.ImageBitmap
 import androidx.lifecycle.Lifecycle
 import androidx.lifecycle.compose.LifecycleEventEffect
 import at.asitplus.catchingUnwrapped
-import at.asitplus.wallet.app.common.DCAPIInvocationData
+import at.asitplus.wallet.app.common.dcapi.DCAPIInvocationData
 import at.asitplus.wallet.app.common.ErrorService
 import at.asitplus.wallet.app.common.WalletDependencyProvider
 import at.asitplus.wallet.app.common.WalletMain
@@ -47,16 +47,11 @@ fun App(walletDependencyProvider: WalletDependencyProvider) {
 
             LifecycleEventEffect(Lifecycle.Event.ON_RESUME) {
                 Napier.d("Lifecycle.Event.ON_RESUME")
-                // TODO is this the best place to sync the credentials with the system?
-                walletMain.updateDigitalCredentialsAPIIntegration()
             }
-        }.onFailure {
-            val errorService: ErrorService = koinInject()
-            errorService.emit(it)
-        }
 
-        WalletTheme {
-            WalletNavigation()
+            WalletTheme {
+                WalletNavigation()
+            }
         }
     }
 }