+++ shared/src/commonMain/kotlin/at/asitplus/wallet/app/common/dcapi/data/preview/PreviewRequest.kt
@@ -1,10 +1,9 @@
-package at.asitplus.wallet.app.common.dcapi
+package at.asitplus.wallet.app.common.dcapi.data.preview
 
 import at.asitplus.catching
 import at.asitplus.wallet.lib.data.vckJsonSerializer
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
-import kotlinx.serialization.encodeToString
 import kotlinx.serialization.json.JsonObject
 
 @Serializable