+++ shared/src/commonMain/kotlin/domain/BuildAuthenticationConsentPageFromAuthenticationRequestUriUseCase.kt
@@ -1,6 +1,7 @@
 package domain
 
 import at.asitplus.KmmResult
+import at.asitplus.dcapi.request.Oid4vpDCAPIRequest
 import at.asitplus.wallet.app.common.PresentationService
 import at.asitplus.wallet.lib.data.vckJsonSerializer
 import io.github.aakira.napier.Napier
@@ -9,21 +10,33 @@ import ui.navigation.routes.AuthenticationViewRoute
 class BuildAuthenticationConsentPageFromAuthenticationRequestUriUseCase(
     val presentationService: PresentationService,
 ) {
-    suspend operator fun invoke(requestUri: String): KmmResult<AuthenticationViewRoute> {
-        val request = presentationService.parseAuthenticationRequestParameters(requestUri).getOrElse {
+    suspend operator fun invoke(
+        requestUri: String,
+        incomingDcApiRequest: Oid4vpDCAPIRequest? = null
+    ): KmmResult<AuthenticationViewRoute> {
+        val request = presentationService.parseAuthenticationRequestParameters(
+            requestUri,
+            incomingDcApiRequest
+        ).getOrElse {
             Napier.d("authenticationRequestParameters: $it")
             return KmmResult.failure(it)
         }
 
-        val preparationState = presentationService.startAuthorizationResponsePreparation(request).getOrElse {
-            return KmmResult.failure(it)
-        }
+        val preparationState =
+            presentationService.startAuthorizationResponsePreparation(request).getOrElse {
+                Napier.e("Failure", it)
+                return KmmResult.failure(it)
+            }
 
         // TODO: extract recipient name from the metadataResponse; the data is not yet being delivered though
         return KmmResult.success(
             AuthenticationViewRoute(
-                authenticationRequestParametersFromSerialized = vckJsonSerializer.encodeToString(request),
-                authorizationPreparationStateSerialized = vckJsonSerializer.encodeToString(preparationState),
+                authenticationRequestParametersFromSerialized = vckJsonSerializer.encodeToString(
+                    request
+                ),
+                authorizationPreparationStateSerialized = vckJsonSerializer.encodeToString(
+                    preparationState
+                ),
                 recipientLocation = request.parameters.clientId ?: "",
                 isCrossDeviceFlow = false,
             )