+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Release 5.7.0
  * Require biometric auth to use holder key
  * Update to VC-K 5.7.0, adding encryption and key agreement for all targets
+ * Add support for OID4VP and ISO 18013-7 Annex C over the Digital Credentials API
 
 # Release 5.6.5
  * Credentials: Safe decoding of images in credentials